python_programming_task/proxy/requirements.txt
#required
pyjwt==2.1.0
aiohttp==3.7.4.post0